Reorganization of API and implementations to prepare for RSA
I have detected a layer between API and impl which consist of the APIs of some implemented algorithms. For example: ECC and RSA have specific domain classes which can be considered as part of their API, but I don't want to include them in cipher's API. Same goes for ufixnum library inside cipher implementing Uint8, Uint16, etc. Thus, I have reorganized some libraries. Hopefully this will make things more maintainable when cipher arrives at 1.0.

+++ lib/_parts/api/random.dart
@@ -29,13 +29,13 @@ abstract class SecureRandom {
   void seed( CipherParameters params );
 
   /// Get one byte long random int.
-  Uint8 nextUint8();
+  int nextUint8();
 
   /// Get two bytes long random int.
-  Uint16 nextUint16();
+  int nextUint16();
 
   /// Get four bytes long random int.
-  Uint32 nextUint32();
+  int nextUint32();
 
   /// Get a random [BigInteger] of [bitLength] bits.
   BigInteger nextBigInteger( int bitLength );
